requirements.txt
PyQt5>=5.15.0
ffmpeg-python>=0.2.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
ollama>=0.1.0
librosa>=0.10.0
pydub>=0.25.1
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0
requests>=2.31.0
python-dateutil>=2.8.0